Code quality improvements surfaced by IntelliJ.

+++ solarnet/datum/src/main/java/net/solarnetwork/central/datum/imp/biz/dao/DaoDatumImportBiz.java
@@ -252,7 +252,7 @@ public class DaoDatumImportBiz extends BaseDatumImportBiz
 			if ( this.resourceStorageWaitMs > 0 ) {
 				try {
 					Boolean result = saveFuture.get(this.resourceStorageWaitMs, TimeUnit.MILLISECONDS);
-					if ( result != null && !result.booleanValue() ) {
+					if ( result != null && !result ) {
 						throw new RuntimeException("Failed to save resource.");
 					}
 				} catch ( TimeoutException e ) {
@@ -265,7 +265,7 @@ public class DaoDatumImportBiz extends BaseDatumImportBiz
 							this.resourceStorageWaitMs, f.getName(), rss.getUid());
 				} catch ( ExecutionException | RuntimeException e ) {
 					log.error("Error saving data import file [{}] to resource storage [{}]: {}",
-							this.resourceStorageWaitMs, f.getName(), rss.getUid(), e.getCause(), e);
+							f.getName(), rss.getUid(), e.getCause(), e);
 					Throwable t = e;
 					while ( t.getCause() != null ) {
 						t = t.getCause();
@@ -295,8 +295,9 @@ public class DaoDatumImportBiz extends BaseDatumImportBiz
 			final int prev = logCount.get();
 			final int diff = now - prev;
 			if ( diff > 10 || now >= 100 ) {
-				String msg = String.format("Saved %.1f%% of datum import [%s] to resource storage [%s]",
-						percent, f.getName(), rss.getUid());
+				String msg = String.format(
+						"Saved %.1f%% of user %d datum import %s [%s] to resource storage [%s]", percent,
+						id.getUserId(), id.getId(), f.getName(), rss.getUid());
 				log.info(msg);
 				logCount.set(now);
 			}
@@ -389,7 +390,7 @@ public class DaoDatumImportBiz extends BaseDatumImportBiz
 	@Override
 	public Collection<DatumImportStatus> datumImportJobStatusesForUser(Long userId,
 			Set<DatumImportState> states) {
-		return jobInfoDao.findForUser(userId, states).stream().map(d -> taskForJobInfo(d))
+		return jobInfoDao.findForUser(userId, states).stream().map(this::taskForJobInfo)
 				.collect(toList());
 	}
 
@@ -417,10 +418,12 @@ public class DaoDatumImportBiz extends BaseDatumImportBiz
 
 	@Override
 	public Collection<DatumImportStatus> deleteDatumImportJobsForUser(Long userId, Set<String> jobIds) {
+		if ( jobIds == null || jobIds.isEmpty() ) {
+			return Collections.emptyList();
+		}
 		Set<DatumImportState> allowStates = EnumSet
 				.complementOf(EnumSet.of(DatumImportState.Claimed, DatumImportState.Executing));
-		Set<UUID> ids = (jobIds != null ? jobIds.stream().map(s -> UUID.fromString(s)).collect(toSet())
-				: null);
+		Set<UUID> ids = jobIds.stream().map(UUID::fromString).collect(toSet());
 		int deleted = jobInfoDao.deleteForUser(userId, ids, allowStates);
 		log.debug("Deleted {} import jobs for user {} matching ids {} in states {}", deleted, userId,
 				jobIds, allowStates);
@@ -429,9 +432,9 @@ public class DaoDatumImportBiz extends BaseDatumImportBiz
 			return userId.equals(task.getUserId()) && jobIds.contains(task.getJobId())
 					&& allowStates.contains(task.getJobState());
 		});
-		return jobInfoDao.findForUser(userId, null).stream().filter(job -> {
-			return userId.equals(job.getUserId()) && jobIds.contains(job.getId().getId().toString());
-		}).map(job -> taskForJobInfo(job)).collect(toList());
+		return jobInfoDao.findForUser(userId, null).stream().filter(
+				job -> userId.equals(job.getUserId()) && jobIds.contains(job.getId().getId().toString()))
+				.map(this::taskForJobInfo).collect(toList());
 	}
 
 	private ImportContext createImportContext(DatumImportJobInfo info,
@@ -491,7 +494,7 @@ public class DaoDatumImportBiz extends BaseDatumImportBiz
 					t = t.getCause();
 				}
 				log.error("Error fetching data import file [{}] from resource storage [{}]: {}",
-						this.resourceStorageWaitMs, dataFile.getName(), rss.getUid(), t.getMessage(), t);
+						dataFile.getName(), rss.getUid(), t.getMessage(), t);
 				throw new RuntimeException(
 						"Failed to save import data to resource storage: " + t.getMessage());
 			}
@@ -504,8 +507,7 @@ public class DaoDatumImportBiz extends BaseDatumImportBiz
 		final ResourceStorageService rss = this.resourceStorageService;
 		if ( rss != null && rss.isConfigured() ) {
 			try {
-				CompletableFuture<Set<String>> f = rss
-						.deleteResources(Arrays.asList(dataFile.getName()));
+				CompletableFuture<Set<String>> f = rss.deleteResources(List.of(dataFile.getName()));
 				f.get(resourceStorageWaitMs, TimeUnit.MILLISECONDS);
 				log.info("Deleted completed data import resource from storage [{}]: {}", rss.getUid(),
 						dataFile.getName());
@@ -523,7 +525,7 @@ public class DaoDatumImportBiz extends BaseDatumImportBiz
 					t = t.getCause();
 				}
 				log.error("Error deleting data import file [{}] from resource storage [{}]: {}",
-						this.resourceStorageWaitMs, dataFile.getName(), rss.getUid(), t.getMessage(), t);
+						dataFile.getName(), rss.getUid(), t.getMessage(), t);
 			}
 		}
 	}
@@ -564,20 +566,16 @@ public class DaoDatumImportBiz extends BaseDatumImportBiz
 					.ownershipsForUserId(info.getUserId());
 			final SecurityPolicy policy = tokenPolicyForId(info.getTokenId());
 
-			final Set<Long> allowedNodeIds;
-			final Map<Long, ZoneId> tzMap;
+			final Map<Long, ZoneId> tzMap; // key set represents the allowed node IDs as well
 			if ( ownerships != null ) {
-				allowedNodeIds = new HashSet<>(ownerships.length);
 				tzMap = new HashMap<>(ownerships.length);
 				for ( SolarNodeOwnership ownership : ownerships ) {
 					if ( policy == null || policy.getNodeIds() == null || policy.getNodeIds().isEmpty()
 							|| policy.getNodeIds().contains(ownership.getNodeId()) ) {
-						allowedNodeIds.add(ownership.getNodeId());
 						tzMap.put(ownership.getNodeId(), ownership.getZone());
 					}
 				}
 			} else {
-				allowedNodeIds = Collections.emptySet();
 				tzMap = Collections.emptyMap();
 			}
 
@@ -676,8 +674,7 @@ public class DaoDatumImportBiz extends BaseDatumImportBiz
 					root = root.getCause();
 				}
 				StringBuilder msg = new StringBuilder();
-				if ( e instanceof AuthorizationException ) {
-					AuthorizationException ae = (AuthorizationException) e;
+				if ( e instanceof AuthorizationException ae ) {
 					if ( ae.getReason() == Reason.ACCESS_DENIED ) {
 						msg.append("Not authorized to load data for ")
 								.append(ae.getId() instanceof Long ? "node" : "source").append(" ")
@@ -693,8 +690,7 @@ public class DaoDatumImportBiz extends BaseDatumImportBiz
 					} else {
 						msg.append(root.getMessage());
 					}
-					if ( e instanceof DatumImportValidationException ) {
-						DatumImportValidationException dive = (DatumImportValidationException) e;
+					if ( e instanceof DatumImportValidationException dive ) {
 						if ( root.getMessage() != null ) {
 							msg.insert(0, " ");
 							msg.insert(0, dive.getMessage());
@@ -752,8 +748,7 @@ public class DaoDatumImportBiz extends BaseDatumImportBiz
 
 		@Override
 		public void handleLoadingException(Throwable t, LoadingContext<GeneralNodeDatum> context) {
-			if ( t instanceof DatumImportValidationException ) {
-				DatumImportValidationException ve = (DatumImportValidationException) t;
+			if ( t instanceof DatumImportValidationException ve ) {
 				throw new DatumImportException(ve.getMessage(), ve.getCause(), ve.getLineNumber(),
 						ve.getLine(), context.getCommittedCount());
 			}
